tests/wpt/web-platform-tests/css/filter-effects/parsing/color-interpolation-filters-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>Filter Effects Module Level 1: getComputedValue().colorInterpolationFilters</title>
<link rel="help" href="https://drafts.fxtf.org/filter-effects/#ColorInterpolationFiltersProperty">
<meta name="assert" content="color-interpolation-filters computed value is as specified.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("color-interpolation-filters", "auto");
test_computed_value("color-interpolation-filters", "srgb");
test_computed_value("color-interpolation-filters", "linearrgb");
</script>
</body>
</html>
