LayoutTests/imported/w3c/web-platform-tests/css/filter-effects/parsing/color-interpolation-filters-computed-expected.txt


PASS Property color-interpolation-filters value 'auto'
FAIL Property color-interpolation-filters value 'srgb' assert_equals: expected "srgb" but got "sRGB"
FAIL Property color-interpolation-filters value 'linearrgb' assert_equals: expected "linearrgb" but got "linearRGB"
